requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
flask==2.3.3
pillow==10.0.0
scikit-learn==1.3.0
sqlite3